pyproject.toml
[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[project]
name = "imha"
version = "0.1.0"
description = "A more constrained and friendlier fork of ImageHash."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=9.1.0",
]

[dependency-groups]
dev = [
    "black",
    "mypy",
    "pre-commit",
    "pytest",
    "ruff",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff.lint]
select = [
    "A",
    "ANN",
    "ARG",
    "ASYNC",
    "B",
    "BLE",
    "C4",
    "D",
    "DTZ",
    "E",
    "EM",
    "EXE",
    "F",
    "FBT",
    "FIX",
    "FLY",
    "FURB",
    "G",
    "I",
    "ICN",
    "INP",
    "ISC",
    "LOG",
    "N",
    "PERF",
    "PGH",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "RET",
    "RSE",
    "RUF",
    "S",
    "SIM",
    "SLF",
    "SLOT",
    "T10",
    "T20",
    "TC",
    "TID",
    "TRY",
    "UP",
    "W",
    "YTT",
]

future-annotations = true

ignore = [
    "ANN401",   # We want to be able to annotate with Any.
    "PLR2004",  # We want to be able to use magic values.
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "ANN",      # We want to avoid type annotations.
    "D",        # We want to avoid docstrings.
    "INP",      # We want tests to be an implicit namespace package.
    "S101",     # We want to use assert.
]

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"numpy".msg = "Use the Python standard library instead."

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.lint.pycodestyle]
max-doc-length = 72

[tool.ruff.lint.pydocstyle]
convention = "pep257"
